backend/requirements.txt
fastapi==0.115.4
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
mistralai==1.5.1
python-multipart==0.0.20
httpx==0.28.1
groq==0.18.0
requests==2.32.3
pymongo==4.11.2
tqdm==4.67.1
cloudinary==1.42.2
pymupdf==1.25.3